refactor: Separate code manipulating assets
