+++ src/output/webbook.ts
@@ -18,7 +18,7 @@ import {
   getIgnoreAssetPatterns,
   getIgnoreThemeExamplePatterns,
   globAssetFiles,
-} from '../processor/compile.js';
+} from '../processor/asset.js';
 import {
   createVirtualConsole,
   fetchLinkedPublicationManifest,
@@ -405,6 +405,7 @@ export async function copyWebPublicationAssets({
   outputDir: string;
   manifestPath: string;
 }): Promise<{ manifest: PublicationManifest; actualManifestPath: string }> {
+  debugger;
   const relExportAliases = exportAliases
     .map(({ source, target }) => ({
       source: upath.relative(input, source),
